[RORDEV-1636] Give every ES|QL type a file of its own or an owner
Three files were holding whatever shared an import list. Each concept now has one
file, named after it, declaring the trait or class and its companion - so where a
type lives says what it is, and nothing has to be hunted for.

The two exceptions are the ones with a real owner: a span and a source location mean
nothing without the query they point into, so they moved into Query's companion.
